requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bitrix==0.0.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
coverage==7.5.3
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
Faker==25.8.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
msgpack==1.0.8
phonenumbers==8.13.37
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.4.post2
